… go/types/sequence_iterator_test.go
… go/types/sequence_readahead.go
… go/util/math/minmax.go