This change implements read-ahead in the sequence_cursor to speed up sequence reads.

Some before and after numbers reading sf-registered-businesses/raw:

| Origin | Before | After |
| --- | --- | --- |
| ldb | 45 MB/s | 60 MB/s |
| localhost | 7.5 MB/s | 27 MB/s |
| demo.noms.io | 140 KB/s | 2.1 MB/s |

The approach:

For each meta-sequence that contains leaf sequences, start reading ahead in
parallel and deliver in order to a buffered channel. Each advance of the cursor gets
the next sequence in the read-ahead channel.

Since this is operating on the sequence_cursor, it benefits all sequence types.
## toward: #2079