@ehalpern 2016-10-18
Thanks for your reviews @cmasone-attic & @arv. I've tried address all the small issues.

I need to think a bit more about how to best address the problem of more tightly constraining cursors used for read ahead. I wanted to first build a solution that would work with minimal API changes to prove it out. API changes are going to be a little hairy since the sequence_cursor is used from many call sights and it's current abstraction is a little leaky.

CC: @kalman & @rafael-atticlabs