@ehalpern 2016-10-19
PTAL @rafael-atticlabs, @kalman
cc: @cmasone-attic, @arv 

Checkpointing the review comments, there's one remaining issue: what is the best way to manage the background read-ahead work? 

**Current approach:**

 - Implement read-ahead in the sequence cursor, limiting the read-ahead to leaf chunks
 - Whenever the cursor moves to a new metaSequence chunk (that references leaf chunks), 
   it initiates read-ahead by creating an orderedparallel goroutine pool whose input channel is 
   leaf chunk idx's and output channel is (chunk idx, chunk byte) pairs. 
 - When the cursor moves, it checks the output channel to see if the chunk it's looking for is in
   the queue. 

Trade-offs:
  - + Resource usage is bounded by the size of the goroutine pool
  - + Works well if cursor is scanning forward
  - - Tricky to handle cases where cursor isn't moving forward
  - - Tricky to managing the lifetime of the goroutine pool. To be robust, this requires 
       implementing a close() method on the cursor to ensure the goroutines are always 
       terminated. If we go this route, I'd propose introducing a new, higher level type for 
       forward iteration (sequenceIterator? or sequenceReader?)  that includes a Close()
       method. 

** Pool-less approach **

@kalman suggested in this thread and subsequent discussion considering a solution that 
doesn't require a goroutine pool. The simplest example of this would be just reading ahead 
one chunk by starting a new goroutine on each advance. Here the goroutine could read the 
chunk, deliver it on the read ahead channel and terminate. This way, goroutines are always short lived and don't require termination. 

This is an elegant idea, but I don't think there's a good way to extend it to read-ahead by more than
one chunk if we continue to deliver the read-ahead chunks on a channel. The problem is that as soon as you introduce more that one goroutine working in parallel, you need a mechanism to order the read-ahead channel (so that chunks are delivered in index order). To do this requires introducing at least one long running goroutine that reads from the unordered channel and and writes to an ordered channel. 

It's possible to handle the ordering problem without a long-lived goroutine, but I can't yet think of a solution that isn't significantly more complicated.

** Read-thru cache approach **

A third approach would be do the caching at a lower level. Here's the idea: 
 - Ensure that the ChunkStore is always fronted by ReadThruStore
 - In the sequence cursor, initiate n read-ahead by starting n goroutines to read n chunks ahead 
 - Each goroutine will just simply read a chunk. By doing so, it will make it likely that the next 
   read of the chunk will find it in the CachingChunkStore into the underlying cache.
 - On each advance, start one more goroutine to read the next chunk into the cache.

Trade-offs:
 - + Simple solution that requires very little code
 - + Doesn't require managing long-lived goroutines
 - - Requires always using CachingChunkStore to get read-ahead performance

** Next steps **

 - The ReadThruStore approach seems very attractive on it's face. It's pool-less and it's 
    simple. Is it practical to enable the the store cache by default? 
 - If the ReadThruStore approach isn't viable, I'm leaning toward improving on the 
   current approach rather than implementing (what seems like) a more complicated 
   pool-less approach. My sense is that a more constrained api for forward scanning is a 
   good direction regardless of the read-ahead implementation; and that once that exists, 
   including a Close method is not too onerous.

Look forward to your feedback.