@kalman 2016-10-19
Nice writeup.

Could there be a 4th approach, which uses the same idea as 3 but implementing the cache on the sequenceCursor itself? A sort of hybrid between 2 and 3.

Very simplistic strawman, and without an attempt to deal with tricky issues (like overlowing into the next chunk, retreating, initialization, etc), instead of:
```go
func (c *sequenceCursor) getChildSequence() sequence {
  return c.seq.getChildSequence(c.idx)
}
```
give the cursor a map of index -> channel, like:
```go
type sequenceCursor struct {
  ...
  raCache map[int]chan sequence
}
```
and implement getChildSequence like (simplistically):
```go
func (c *sequenceCursor) getChildSequence() sequence {
  // Set up next readahead.
  raIdx = c.idx + raSize
  raChan = make(chan sequence)
  c.raCache[raIdx] = raChan
  go func() {
    raChan <- c.seq.getChildSequence(raIdx)
  }()

  // Fetch the current child sequence from the readahead cache.
  return <-c.seq.raCache[c.idx]
}
```

I also agree that either way we'll want a cursor-esque API that doesn't suffer from really subtle issues like this.