@ehalpern 2016-10-19
Thanks @kalam: Yes, I think a hybrid is possible and your sketch looks like a good start. Your use of channels in the map is an interesting simplification over storing the actual sequence there. Makes me realize that I've only scratched the surface of what can be done with them.

I'm still keen on the idea of a shared chunk cache if practical. It simplifies this solution further by better separating concerns and it may have more far reaching performance advantages. 

If not practical, however, your suggestion seems like a good direction to go.