@kalman 2016-10-19
Makes sense. I think @cmasone-attic and @rafael-atticlabs have thought much more about caching at the chunk store layer than I have, I'll let them speak for that.