@ehalpern 2016-11-02
PTAL @rafael-atticlabs: 
FYI @kalman, @cmasone-attic 

I've changed the read-ahead mechanism to use short-lived go routines and unbuffered channels based on @kalman's suggestion. This removes the need to close at the end of iteration.

I'm uncertain whether to proceed with restricting read-ahead to a more limited iteration API ( sequenceCursor.iter++) or keeping it enabled for all users of sequence_cursor. 

A new, more restricted API is the clean thing to do, but it will require some surgery since there are many call sites that need to change. 

Enabling read-ahead for all callers of sequence cursor (or enabling it by a switch) has the pragmatic benefit that it works in all cases where it's needed without any code changes. It has the drawbacks that it (1) wastes resources in the cases where it's not needed (which are not numerous) and (2) may hinder more aggressive optimizations.