@ehalpern 2016-11-10
PTAL @rafael-atticlabs 
cc: @kalman 

I've reworked read-ahead to use a cursor above the leaves so that it isn't limited to the current chunk. The performance diff is significant:

| Origin | Before | One chunk at a time | Across chunks |
| --- | --- | --- |---|
| ldb | 45 MB/s | 60 MB/s | 60 MB/s |
| localhost | 7.5 MB/s | 27 MB/s | 38 MB/s |
| demo.noms.io | 140 KB/s | 2.1 MB/s | NA* |

* demo.noms.io has inconsistent throughput from office 

I compared the shorted-lived go-routine approach to the long-lived pool based approach and found that short-lived approach to have comparable performance. This result combined with the big advantage of not requiring the caller to shut down the threadpool (with a close) convinced me to take the first approach. 

I've also creating a new sequenceIterator class which encapsulates forward iteration. It's currently only used for BlobReader and sequenceCursor.iter, but I'd recommend we refactor all iterators to use this as their foundation. This reduces code and reduces the surface of sequence_iterator.